:white_check_mark: Bootstrap tests, ++numBrowsers
